Commit for day 2024-01-08
Today, I hit a major roadblock for figuring out part two, which is that
it seems I know much less about the system than I initially thought. I
thought that every rock was hit at some "t+1", "t+2" and so on, but it
turns out they can be hit at any point in the future, this makes things
much more challenging.

I am not yet sure how I will solve this, but one thing which is nice,
which I just realized, is that we can solve this for the 3 axis
seperatly, being that the solution is the intersection between the
possible solutions for the x, y and z planes. I'll try to figure out if
there is something more to do here.

I'll spend some more time on this tomorrow, but if I cannot crack it
then, I'll take a look at reddit to see if there is anything I can get
out of that.

My current best clue, is to look at the way the stones are moving each
turn, and I think it's established that the stone cannot move slower
than the stone which is the furthers from it in that direction when it
starts. But I really don't know how this is supposed to be solved.

I tried converting it into a set of linear equations, but I realized
after a while that it's impossible to do that the way I wanted, as we
end up with a "s_1 * v_x" step, where both of them are variables. Maybe
there is some other, cleverer way to do this, but for now, I'll have to
look into some other angles.