Cleanup of 2023 Day 17
So this was an interesting day and even if I'm committing this now, I
did the changes the day it was solved, but I didn't have time to write a
longer commit message. But I remember this day quite well, so I'm
writing them now.

So it turns out I was extremely lucky today. I lucked out on both parts,
in two different ways. The first mistake I made, which I made in both
parts, is that I started the "going right" starting point, with 1 step,
which is not correct. I did however start the "going down" with 0 steps,
which is the correct thing to do. And luckily both part one and part
two's solutions start with going down.

The second mistake I made was that I started the left and right steps
with 0 steps, instead of 1. This just had the effect of making it so I
had to do less than or equal to 2 instead of 3, but I wondered why it
was like this during the competition.

The last piece of luck I had was that I didn't read the final line, that
you had to move at least 3 squares, also on the last move! In the end it
didn't matter for my solution, but it easily could have!

The way I figured out most of these mistakes, was that I actually
thought I had found a bug in the puzzle. I got a path that was cheaper
than the answer on part two, and wondered how this could be. This was
after I fixed the "1 step on going right" bug I describe above. After
visualizing the mistake and other things, I made sure that the route
described was correct and went to reddit to make a comment. That's when
I read in the mega thread about the fact that you had to move at least 3
squares also on the last move and also the fact that other inputs had
been tricky with the starting with 0 steps policy.

Overall I was very lucky on this day, I think it could have taken a long
time to debug why my answer was not the answer, especially as it was
lower than the actual answer.