Optimize 2023 Day 24 Part 2
So I had to look at reddit to get this done, but now that I have, I see
that there really is many ways to do this, but it can be done very
straightforward with just 3 hailstones. I'll write the full derivation
of this in the comments for the solution, but for now I just wanted to
get a comment in.