2023 Day 21 Part two
And with this, I have finally solved all the tasks in 2023. I was
reluctant to enter into this one, because I thought it would get rather
ugly, but once again, the input is very kind.

I was able to solve it, by noticing that there is no stones in the
starting row or the starting column. There are also no stones on the
edge of the map.

Using this, I partition the grid into 8 zones and I calculate how many
plots we get to for each zone. This is fast and runs in O(B) where B is
the size of the initial board, as we have to do a BFS search on the
board to figure out which spots are reachable and the cost of reaching
each space.

reading the reddit thread, it appears that it's possible to solve this
in some other way, using a diamond shape. I slept really badly tonight
so I'll look into that a bit later.