Solve 2016 Day 22 Part 2
This turned out to be very easy once you realized that we are not trying
to solve the generalized version of the problem at all, but rather
something specific to our input. It can be reduced to just a few types
of nodes, either too big, empty or normal nodes. After that it's just a
simple calculation of finding the shortest path to the tile before the
target data and calculating how many steps it would take to move this
data all the way to the left.

This is of course assuming, that we have no blockers in the top column.